seastore/rbm: rename nvme_device to random_block_device
rename nvme_device namespace to random_block_device, and
introduce random_block_device::nvme
