+++ src/crimson/os/seastore/journal.cc
@@ -13,7 +13,7 @@ JournalRef make_segmented(SegmentProvider &provider)
 }
 
 JournalRef make_circularbounded(
-  crimson::os::seastore::nvme_device::RBMDevice* device,
+  crimson::os::seastore::random_block_device::RBMDevice* device,
   std::string path)
 {
   return std::make_unique<CircularBoundedJournal>(device, path);
